refactor mod structure
