+++ crates/swc_ecma_utils/src/function/mod.rs
@@ -1,0 +1,5 @@
+mod fn_env_hoister;
+mod function_wrapper;
+
+pub use fn_env_hoister::*;
+pub use function_wrapper::*;
